tools/can-monitor/requirements.txt
PyQt5>=5.15
pyqtgraph>=0.13
pyserial>=3.5
cantools>=39.0
